+++ comicbox/box/page_filenames.py
@@ -1,12 +1,14 @@
 """Calculate page filenames."""
 
-from pathlib import Path
+import re
 from sys import maxsize
 
 from comicbox.box.archive import archive_close
 from comicbox.box.archive_read import ComicboxArchiveReadMixin
 
-IGNORE_PREFIXES = (".", "__macosx")
+# ignore dotfiles but not relative ../ leaders.
+# ignore macos resource forks
+_IGNORE_RE = re.compile(r"(?:^|\/)(?:\.[^\.]|__MACOSX)", re.IGNORECASE)
 
 
 class ComicboxPageFilenamesMixin(ComicboxArchiveReadMixin):
@@ -24,11 +26,9 @@ class ComicboxPageFilenamesMixin(ComicboxArchiveReadMixin):
         else:
             self._page_filenames = []
             for filename in archive_filenames:
-                for part in Path(filename).parts:
-                    for prefix in IGNORE_PREFIXES:
-                        if part.lower().startswith(prefix):
-                            continue
-                if self.IMAGE_EXT_RE.search(filename) is not None:
+                if not _IGNORE_RE.search(filename) and self.IMAGE_EXT_RE.search(
+                    filename
+                ):
                     self._page_filenames.append(filename)
 
     def get_page_filenames(self):